wip(multitable): extract runShortcutCore to ai-bulk-shared (inline path identical)
